Renamed LoginData to LoginForm
