+++ app/views/user/login.scala.html
@@ -1,4 +1,4 @@
-@(loginForm: Form[viewmodels.LoginData])
+@(loginForm: Form[viewmodels.user.LoginForm])
 @import helper._
 
 @main("BringToAfrica") {
